Ignore missing broker images
There was a bug with loading the remote logos and so I stripped out the
ability to load remote logos without realising that we display brokers
in the broker center that we do not actually support. That means they
don't have a local logo and this was causing the code to crash.

This fix just doesn't set a logo if there is none in the bundle.

In the broker center it will just display the broker name.

The goal is to fix this properly in the new SDK.